+++ app/src/main/java/com/app/arcabyolimpo/data/remote/dto/user/updateuser/UpdateUserDto.kt
@@ -10,10 +10,10 @@ data class UpdateUserDto(
     val roleId: String,
 
     @SerializedName("externalCollab_name")
-    val name: String,
+    val name: String?,
 
     @SerializedName("externalCollab_lastName")
-    val lastName: String,
+    val lastName: String?,
 
     @SerializedName("externalCollab_secondLastName")
     val secondLastName: String,
@@ -25,10 +25,10 @@ data class UpdateUserDto(
     val degree: String,
 
     @SerializedName("externalCollab_email")
-    val email: String,
+    val email: String?,
 
     @SerializedName("externalCollab_phone")
-    val phone: String,
+    val phone: String?,
 
     @SerializedName("externalCollab_status")
     val status: Int,