+++ app/src/main/java/com/app/arcabyolimpo/presentation/screens/beneficiary/AddNewBeneficiary.kt
@@ -60,7 +60,6 @@ fun AddNewBeneficiaryScreen(
                     )
                 }
             },
-            bottomBar = { NavBar() }
         ) { padding ->
             Column(
                 modifier = Modifier