+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Fixed
 - Fix beneficiary list navbar
 - Fix workshops date input
+- Fix Workshops navbar
+- Fix beneficiary registration navBar appearing
 
 ## [0.1.0] - 19/11/2025
 ### Added